requirements.txt
fastapi>=0.111
uvicorn>=0.30
websockets>=12.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
authlib>=1.2.0
itsdangerous>=2.0.0
passlib>=1.7.4
bcrypt>=4.0.1
python-jose>=3.3.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
fastapi-mail>=1.4.0
python-dotenv>=1.0.0
google-genai>=1.33.0
python-multipart>=0.0.6
starlette>=0.36.0
PyJWT>=2.8.0